// wework-api/src/core/models/bag-recipe-detail.ts
export class BagRecipeDetailModel {
	id?: number;
	serialCode?: string;
	monthRecipeBag?: string;
	maquiladorBag?: any;
	numberBags?: number;
	operatingExpense?: number;
	maquila?: number;
	tax?: number;
	commission?: number;
	surcharges?: any;
	maquiladors?: any;
	products?: any;
	payments?: any;
}